Simplify Mount Location
In production environments, we can't depend on Application configs and
environment variables to specify a configuration value for a dependency.

This commit simplifies how we configure the path to Teal in the elixir
and JS code.

Previously we had to set:

* a config value for `base_url` that included both the path and full
origin
* a namespace and scope in the router
* a fixed value in the `ExTeal.Configuration` that fixed the asset
location to `/teal`

With this commit, we only have to specify:

```elixir
  # In the App Router

  scope path: "/teal" do
    forward("/", ExTeal.Router, namespace: "teal")
  end
```

For a default installation where Teal is available at the path `/teal`.

To configure this path, you can now define:

```elixir
  # In the App Router

  scope path: "/admin" do
    forward("/", ExTeal.Router, namespace: "admin")
  end
```

and in the teal manifest:

```
  def path, do: "/admin"
```