DateTime and Date rewrite
This MR replaces `date-fns` with the new library `luxon` from the
Moment team.  It allows us to use easy Intl features and integrates with
the new datetime field to provide localized datetimes.

Depreciates the `Date.with_options/2` function in favor of more explicit
`format/2` and `picker_format/2` functions.

`format/2` accepts an atom that maps to one of the valid localized
signatures of `Luxon` and `picker_format` accepts a string that is used
by flatpickr to format the value displayed in the input field.