Add Type Argument to `apply_options_for/2-3`
Adds a `type` argument to the `apply_options_for` callback on a field.
By passing this value down through the serializers, we can give fields
the ability to manipulate their options based on how the field is being
serialized.

This will be a *breaking* change, required for work on
`ex_teal_dependency_container`