Closes #6, Fixes broken HasOne Index View
The `HasOne` field depends on a `has_one_id` and `has_one_relationship`
as field options to build the Index link and the Detail View Table.

The `has_one_id` field was not being returned in the options, making the
link return an invalid url.

This commit also fixes a bug with the `HasOne` Index field that also
used the `has_one_relationship` and not the relationships' `uri`.

In short, this PR turns the link on an `Index` field where a User
`has_one` Post from `/resources/post/NAN` to `/resources/posts/1`.