fix: Broken tests due to timezone offsets
This commit fixes a bug in the test suite that was caused by looking up
the offsets for a specific timezone by the name of the timezone using
the current time as the time to calculate the offset.  This means that
for timezones that are dst observing, they were returning an incorrect
offset breaking the test suite.