feat: BelongsTo should always be searchable
We ran into some strange performance projects on a client project that
was later identified as a non-searchable belongs to field timing out
requests for update and creation fields by attempting to serialize the
entire contents of a table with hundreds of thousands of rows in it.
There are very few use cases where a belongs to with a fixed, small
collection of potential relationships is desireable.  This commit
removes that default and adds a warning to the `BelongsTo.searchable`
function.