feat: Select fields can represent Ecto.Enums
What changed?
=============

We introduce functionality that allows a `ExTeal.Fields.Select` to
represent a `Ecto.Enum` field.

Why was this change made?
=========================

One of the primary values of `Ecto.Enum` is that it allows you to keep
atom values in a field of a schema.  Previously, values stored as atoms
could never be correctly recalled in a Teal select field because it
always selected the current value from the list of options initated in
`Select.options/2`.

These changes allow for the Ecto.Enum to become the source of the
possible options for the select field, and ensures that the two always
coincide.