fix: Allow Distinct in `handle_index/2` callbacks
This commit allows `handle_index/2` overrides on resources to invoke the
`distinct: true` handler and _also enable column sorting_.  Previously
the aggregation used to count up the total rows of a resource would fail
because the order_by columns must always appear in the select
statement when using `distinct: true`.  This commit replaces the helper
`Repo.aggregate/3` function with a subquery call that excludes
previously set `preload` and `select` statements.